requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.6
langsmith>=0.0.66
peft>=0.8.0
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
google-generativeai>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# API server
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Testing
pytest>=7.4.0

# Evaluation
scikit-learn>=1.3.0
trl
openai
langchain-google-genai